pyproject.toml
[tool.poetry]
name = "datafort"
version = "0.1.0"
description = "Test for vacantion"
authors = ["Igor Olenchuk"]
license = "none"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.8"
asyncio = "3.4.3"
certifi = "2022.9.24"
charset-normalizer = "2.1.1"
idna = "3.4"
numpy = "1.23.4"
pandas = "1.5.1"
psycopg2-binary = "2.9.5"
python-dateutil = "2.8.2"
python-dotenv = "0.21.0"
pytz = "2022.6"
requests = "2.28.1"
six = "1.16.0"
SQLAlchemy = "1.4.44"
urllib3 = "1.26.12"
jproperties = "^2.1.1"
SQLAlchemy-Utils = "^0.38.2"
cryptography = "^37.0.1"
sanic = "^22.3.1"
mypy = "^0.950"
redis = {extras = ["asyncio"], version = "^4.2.2"}

[tool.poetry.dev-dependencies]
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"